test/keccak-normalize-table: move test to the new table module
